requirements.txt
nltk==3.8.1
pyspellchecker==0.8.1 
spacy==3.8.2